+++ src/pages/authPages/LoginPage.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import HeaderForm from '../components/molecules/headerForm/HeaderForm';
-import Footer from '../components/organisms/footer/Footer';
-import LoginOrSignupForm from '../components/organisms/loginOrSignupForm/LoginOrSignupForm';
-import SocialMediaLoginOrSignup from '../components/organisms/socialMediaLoginSignup/SocialMediaLoginOrSignup';
+import HeaderForm from '../../components/molecules/headerForm/HeaderForm';
+import Footer from '../../components/organisms/footer/Footer';
+import LoginOrSignupForm from '../../components/organisms/loginOrSignupForm/LoginOrSignupForm';
+import SocialMediaLoginOrSignup from '../../components/organisms/socialMediaLoginSignup/SocialMediaLoginOrSignup';
 
 const LoginPage = () => (
   <>
