+++ src/pages/authPages/SignUpPage.jsx
@@ -1,9 +1,9 @@
 import React, { useLayoutEffect } from 'react';
 import { useLocation } from 'react-router-dom';
-import HeaderForm from '../components/molecules/headerForm/HeaderForm';
-import Footer from '../components/organisms/footer/Footer';
-import LoginOrSignupForm from '../components/organisms/loginOrSignupForm/LoginOrSignupForm';
-import SocialMediaLoginOrSignup from '../components/organisms/socialMediaLoginSignup/SocialMediaLoginOrSignup';
+import HeaderForm from '../../components/molecules/headerForm/HeaderForm';
+import Footer from '../../components/organisms/footer/Footer';
+import LoginOrSignupForm from '../../components/organisms/loginOrSignupForm/LoginOrSignupForm';
+import SocialMediaLoginOrSignup from '../../components/organisms/socialMediaLoginSignup/SocialMediaLoginOrSignup';
 
 const SignUpPage = () => {
   const location = useLocation();
