--- src/pages/dashboardPages/DashboardRoutes.jsx
@@ -1,16 +1,0 @@
-import { useState } from 'react';
-import { Outlet } from 'react-router-dom';
-import Sidebar from '../../components/organisms/sidebar/Sidebar';
-
-const DashboardRoutes = () => {
-  const [width, setWidth] = useState(64);
-
-  return (
-    <>
-      <Sidebar setWidth={setWidth} />
-      <Outlet context={width} />
-    </>
-  );
-};
-
-export default DashboardRoutes;
